Crate subdomain_enumerator

Modules

enumerator
A subdomain enumerator. Finds subdomains of a given subdomain by querying a public dataset (virustotal.com).

library_enumerator
Contains code for a subdomain enumerator that finds subdomain by trying subdomains generated with a library of common words.

query

results