Function subdomain_enumerator::library_enumerator::enumerate

pub fn enumerate(
    domain: String, 
    library: String, 
    store: Arc<Mutex<HashSet<String>>>, 
    pool: ThreadPool
)

Takes a domain, a library, a store, and a thread pool. For each word supplied in library, checkes to see whether the word specifies a subdomain of domain. If it does, the found subdomain is added to store. The thread pool is used to concurrently check possible subdomains. This means that this function should be somewhat efficient even with large libraries.

Arguments

'domain' - A String with the domain you want to enumerate.

'library' - A String with a filepath to the word list you want to use for library enumeration.

'store' - An Arc<Mutex<HashSet>> that you want to read the subdomains into.

'pool' - A ThreadPool holding threads you want to use for concurrent enumeration.