Struct subdomain_enumerator::query::Query

pub struct Query { /* fields omitted */ }

Represents a user supplied query, where domains is the list of domains that will have their subdomains enumerated and library is a wordlist supplied for library enumeration.

Fields

'domains' - A Vector of Strings holding the domains to be enumerated

'library' - A String with the filepath of a wordlist for library enumeration. Can be empty if there is no wordlist.

'limit' - A usize holding the maximum number of subdomains to be enumerated for each domain.

'num_domains' - The number of domains to be enumerated. Updated automatically by add_domain() and add_domains().

Methods

impl Query

pub fn new() -> Self

pub fn add_domain(&mut self, domain: String)

Add a domain (as a string) to be enumerated.

pub fn add_domains(&mut self, domains: Vec<String>)

Add multiple domains (as a vector of strings) to be enumerated.

pub fn set_library(&mut self, library: String)

pub fn set_limit(&mut self, limit: usize)

pub fn enumerate(&self) -> Results

Performs subdomain enumeration on all domains held in the domains variable. The found subdomains will be returned collectively in a single Results object. If a library is supplied, then both library enumeration and a database query will be performed. If no library is supplied, then only a database query will be performed.

Examples

Performing subdomain enumeration on a query with no library, and therefore no library enumeration. ''' let mut q = Query::new(); q.add_domains(vec!["facebook.com", "google.com"]; let subdomains = q.enumerate(); subdomains.print_subdomains(); '''

Performing subdomain enumeration on a query with a library added. This means that library enumeration will be performed along with normal database querying. Note that a library is passed as a string holding a path to an external file. ''' let mut q = Query::new(); q.add_domains(vec!["facebook.com", "google.com"]; q.set_library("path_to_library.txt"); let subdomains = q.enumerate(); subdomains.print_subdomains(); '''